amq_client/requirements.txt
stomp.py
python-slugify
